requirements.txt
alembic==1.13.0
blinker==1.7.0
click==8.1.7
colorama==0.4.6
Flask==3.0.0
Flask-Migrate==4.0.5
Flask-Script==2.0.6
Flask-SQLAlchemy==3.1.1
greenlet==3.0.2
itsdangerous==2.1.2
Jinja2==3.1.2
Mako==1.3.0
MarkupSafe==2.1.3
psycopg2-binary==2.9.9
python-dotenv==1.0.0
SQLAlchemy==2.0.23
typing_extensions==4.9.0
Werkzeug==3.0.1
pymysql==1.1.0
cryptography==41.0.7
flask-smorest==0.42.3
Flask-JWT-Extended==4.6.0
Flask-Bcrypt==1.0.1
pytest==7.4.4
Flask-Cors==4.0.0

